requirements.txt
# Core packages
pytest==7.4.4
black==24.1.1
flake8==7.0.0
python-dotenv==1.0.0
requests==2.31.0

# Google Calendar API
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# SMS functionality
twilio==8.12.0

# Scheduling
schedule==1.2.1 